Crate sp_domains

Primitives for domains pallet.

Modules

bundle_producer_election

core_api

extrinsics

merkle_tree

proof_provider_and_verifier

storage

valued_trie

Structs

BlockFees

Bundle

Domain bundle.

BundleDigest

A digest of the bundle

BundleHeader

DomainAllowlistUpdates

Domain chains allowlist updates.

DomainBundleLimit

DomainId

Unique identifier of a domain.

DomainInstanceData

DomainInstanceData is used to construct the genesis storage of domain instance chain

DomainSudoCall

Domain Sudo runtime call.

ExecutionReceipt

Receipt of a domain block execution.

GenesisDomain

InboxedBundle

InboxedBundle represents a bundle that was successfully submitted to the consensus chain

OperatorKey

A type that implements BoundToRuntimeAppPublic, used for operator signing key.

OperatorSigningKeyProofOfOwnershipData

Preimage to verify the proof of ownership of Operator Signing key. Operator owner is used to ensure the signature is used by anyone except the owner of the Signing key pair.

ProofOfElection

RuntimeObject

SealedBundleHeader

Header of bundle.

SealedSingletonReceipt

SingletonReceipt

Singleton receipt submit along when there is a gap between domain_best_number and HeadReceiptNumber

Transfers

Type that holds the transfers(in/out) for a given chain.

Enums

BundleValidity

ChainId

Identifier of a chain.

DomainDigestItem

Domains specific digest item.

ExtrinsicDigest

Represents the extrinsic either as full data or hash of the data.

InvalidBundleType

Bundle invalidity type

InvalidReceipt

Receipt invalidity type.

OperatorAllowList

Type that represents an operator allow list for Domains.

OperatorRewardSource

PermissionedActionAllowedBy

Permissioned actions allowed by either specific accounts or anyone.

ReceiptValidity

RuntimeType

Types of runtime pallet domains currently supports

Constants

DOMAIN_EXTRINSICS_SHUFFLING_SEED_SUBJECT

Extrinsics shuffling seed

DOMAIN_STORAGE_FEE_MULTIPLIER

EMPTY_EXTRINSIC_ROOT

Empty extrinsics root.

INITIAL_DOMAIN_TX_RANGE

Initial tx range = U256::MAX / INITIAL_DOMAIN_TX_RANGE.

KEY_TYPE

Key type for Operator.

Traits

BundleProducerElectionApi

DomainBundleSubmitted

Post hook to know if the domain had bundle submitted in the previous block.

DomainOwner

Trait to check domain owner.

DomainsApi

API necessary for domains pallet.

DomainsDigestItem

Domains specific digest items.

DomainsTransfersTracker

Trait that tracks the balances on Domains.

OnChainRewards

Hook to handle chain rewards.

OnDomainInstantiated

A hook to call after a domain is instantiated

Functions

calculate_max_bundle_weight_and_size

Calculates the max bundle weight and size

derive_domain_block_hash

domain_account_storage_key

Account info on frame_system on Domains

domain_total_issuance_storage_key

Total issuance storage for Domains.

dummy_opaque_bundle

operator_block_fees_final_key

This is a representation of actual Block Fees storage in pallet-block-fees. Any change in key or value there should be changed here accordingly.

self_domain_id_storage_key

The storage key of the SelfDomainId storage item in the pallet-domain-id

signer_in_tx_range

Checks if the signer Id hash is within the tx range

system_digest_final_key

Digest storage key in frame_system. Unfortunately, the digest storage is private and not possible to derive the key from it directly.

Type Aliases

ChannelId

Channel identity.

EpochIndex

Type representing domain epoch.

ExecutionReceiptFor

ExtrinsicsRoot

The Trie root of all extrinsics included in a bundle.

HeaderHashFor

Type alias for Header hash.

HeaderHashingFor

Type alias for Header Hashing.

HeaderNumberFor

Type alias for Header number.

OpaqueBundle

Bundle with opaque extrinsics.

OpaqueBundles

List of OpaqueBundle.

OperatorId

Type representing operator ID

OperatorPair

An operator authority keypair. Necessarily equivalent to the schnorrkel public key used in the main executor module. If that ever changes, then this must, too.

OperatorPublicKey

An operator authority identifier.

OperatorSignature

An operator authority signature.

RuntimeId

Type representing the runtime ID.

StakeWeight

Stake weight in the domain bundle election.