sp_domains::execution_receipt::execution_receipt_v0

Type Alias ExecutionReceiptV0For

pub type ExecutionReceiptV0For<DomainHeader, CBlock, Balance> = ExecutionReceiptV0<NumberFor<CBlock>, BlockHashFor<CBlock>, <DomainHeader as HeaderT>::Number, <DomainHeader as HeaderT>::Hash, Balance>;

Aliased Type

pub struct ExecutionReceiptV0For<DomainHeader, CBlock, Balance> {
    pub domain_block_number: <DomainHeader as Header>::Number,
    pub domain_block_hash: <DomainHeader as Header>::Hash,
    pub domain_block_extrinsic_root: <DomainHeader as Header>::Hash,
    pub parent_domain_block_receipt_hash: <DomainHeader as Header>::Hash,
    pub consensus_block_number: <<CBlock as Block>::Header as Header>::Number,
    pub consensus_block_hash: <CBlock as Block>::Hash,
    pub inboxed_bundles: Vec<InboxedBundle<<DomainHeader as Header>::Hash>>,
    pub final_state_root: <DomainHeader as Header>::Hash,
    pub execution_trace: Vec<<DomainHeader as Header>::Hash>,
    pub execution_trace_root: H256,
    pub block_fees: BlockFees<Balance>,
    pub transfers: Transfers<Balance>,
}

Fields
domain_block_number: <DomainHeader as Header>::Number

The index of the current domain block that forms the basis of this ER.

domain_block_hash: <DomainHeader as Header>::Hash

The block hash corresponding to domain_block_number.

domain_block_extrinsic_root: <DomainHeader as Header>::Hash

Extrinsic root field of the header of domain block referenced by this ER.

parent_domain_block_receipt_hash: <DomainHeader as Header>::Hash

The hash of the ER for the last domain block.

consensus_block_number: <<CBlock as Block>::Header as Header>::Number

A pointer to the consensus block index which contains all of the bundles that were used to derive and order all extrinsics executed by the current domain block for this ER.

consensus_block_hash: <CBlock as Block>::Hash

The block hash corresponding to consensus_block_number.

inboxed_bundles: Vec<InboxedBundle<<DomainHeader as Header>::Hash>>

All the bundles that being included in the consensus block.

final_state_root: <DomainHeader as Header>::Hash

The final state root for the current domain block reflected by this ER.

Used for verifying storage proofs for domains.

execution_trace: Vec<<DomainHeader as Header>::Hash>

List of storage roots collected during the domain block execution.

execution_trace_root: H256

The Merkle root of the execution trace for the current domain block.

Used for verifying fraud proofs.

block_fees: BlockFees<Balance>

Compute and Domain storage fees are shared across operators and Consensus storage fees are given to the consensus block author.

transfers: Transfers<Balance>

List of transfers from this Domain to other chains