sp_domains

Trait DomainBundleSubmitted

pub trait DomainBundleSubmitted {
    // Required method
    fn domain_bundle_submitted(domain_id: DomainId);
}

Post hook to know if the domain had bundle submitted in the previous block.

Required Methods

fn domain_bundle_submitted(domain_id: DomainId)

Called in the next block initialisation if there was a domain bundle in the previous block. This hook if called for domain represents that there is a new domain block for parent consensus block.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.