Type Alias sp_domains::ExecutionReceiptFor

pub type ExecutionReceiptFor<DomainHeader, CBlock, Balance> = ExecutionReceipt<NumberFor<CBlock>, <CBlock as BlockT>::Hash, <DomainHeader as HeaderT>::Number, <DomainHeader as HeaderT>::Hash, Balance>;

Aliased Type

struct ExecutionReceiptFor<DomainHeader, CBlock, Balance> {
    pub domain_block_number: <DomainHeader as Header>::Number,
    pub domain_block_hash: <DomainHeader as Header>::Hash,
    pub domain_block_extrinsic_root: <DomainHeader as Header>::Hash,
    pub parent_domain_block_receipt_hash: <DomainHeader as Header>::Hash,
    pub consensus_block_number: <<CBlock as Block>::Header as Header>::Number,
    pub consensus_block_hash: <CBlock as Block>::Hash,
    pub inboxed_bundles: Vec<InboxedBundle<<DomainHeader as Header>::Hash>, Global>,
    pub final_state_root: <DomainHeader as Header>::Hash,
    pub execution_trace: Vec<<DomainHeader as Header>::Hash, Global>,
    pub execution_trace_root: H256,
    pub total_rewards: Balance,
}

Fields
domain_block_number: <DomainHeader as Header>::Number

The index of the current domain block that forms the basis of this ER.

domain_block_hash: <DomainHeader as Header>::Hash

The block hash corresponding to domain_block_number.

domain_block_extrinsic_root: <DomainHeader as Header>::Hash

Extrinsic root field of the header of domain block referenced by this ER.

parent_domain_block_receipt_hash: <DomainHeader as Header>::Hash

The hash of the ER for the last domain block.

consensus_block_number: <<CBlock as Block>::Header as Header>::Number

A pointer to the consensus block index which contains all of the bundles that were used to derive and order all extrinsics executed by the current domain block for this ER.

consensus_block_hash: <CBlock as Block>::Hash

The block hash corresponding to consensus_block_number.

inboxed_bundles: Vec<InboxedBundle<<DomainHeader as Header>::Hash>, Global>

All the bundles that being included in the consensus block.

final_state_root: <DomainHeader as Header>::Hash

The final state root for the current domain block reflected by this ER.

Used for verifying storage proofs for domains.

execution_trace: Vec<<DomainHeader as Header>::Hash, Global>

List of storage roots collected during the domain block execution.

execution_trace_root: H256

The Merkle root of the execution trace for the current domain block.

Used for verifying fraud proofs.

total_rewards: Balance

All SSC rewards for this ER to be shared across operators.

Implementations

impl<Number, Hash, DomainNumber, DomainHash, Balance> ExecutionReceipt<Number, Hash, DomainNumber, DomainHash, Balance>

pub fn bundles_extrinsics_roots(&self) -> Vec<&DomainHash>

pub fn valid_bundle_digest_at(&self, index: usize) -> Option<DomainHash>where DomainHash: Copy,

pub fn valid_bundle_digests(&self) -> Vec<DomainHash>where DomainHash: Copy,

pub fn valid_bundle_indexes(&self) -> Vec<u32>

impl<Number: Encode + Zero, Hash: Encode + Default, DomainNumber: Encode + Zero, DomainHash: Clone + Encode + Default, Balance: Encode + Zero> ExecutionReceipt<Number, Hash, DomainNumber, DomainHash, Balance>

pub fn hash<DomainHashing: HashT<Output = DomainHash>>(&self) -> DomainHash

Returns the hash of this execution receipt.

pub fn genesis( genesis_state_root: DomainHash, genesis_extrinsic_root: DomainHash, genesis_domain_block_hash: DomainHash ) -> Self

pub fn dummy<DomainHashing>( consensus_block_number: Number, consensus_block_hash: Hash, domain_block_number: DomainNumber, parent_domain_block_receipt_hash: DomainHash ) -> ExecutionReceipt<Number, Hash, DomainNumber, DomainHash, Balance>where DomainHashing: HashT<Output = DomainHash>,