sp_domains::valued_trie

Function valued_ordered_trie_root

pub fn valued_ordered_trie_root<Layout>(
    input: Vec<Value<'_>>,
) -> <<Layout as TrieLayout>::Hash as Hasher>::Out
where
    Layout: TrieLayout,

This is a modified version of trie root that takes trie node values instead of deriving from the actual data. Taken from trie-db as is.

Note: This is an opportunity to push this change upstream but I’m not sure how to present these changes yet. Need to be discussed further.