Crate sp_domains

Primitives for domains pallet.

Modules

bundle_producer_election

domain

Domain-related runtime interface

fraud_proof

merkle_tree

transaction

Structs

Bundle

Domain bundle.

BundleHeader

DomainBlockLimit

DomainId

Unique identifier of a domain.

DomainInstanceData

DomainInstanceData is used to construct RuntimeGenesisConfig which will be further used to construct the genesis block

ExecutionReceipt

Receipt of a domain block execution.

GenesisDomain

GenesisReceiptExtension

A domain genesis receipt extension.

OperatorKey

A type that implements BoundToRuntimeAppPublic, used for operator signing key.

ProofOfElection

SealedBundleHeader

Header of bundle.

Enums

DomainDigestItem

Domains specific digest item.

DomainsFreezeIdentifier

Domains specific Identifier for Balances freeze.

RuntimeType

Types of runtime pallet domains currently supports

Traits

BundleProducerElectionApi

DomainsApi

API necessary for domains pallet.

DomainsDigestItem

Domains specific digest items.

GenerateGenesisStateRoot

Functions

signer_in_tx_range

Checks if the signer Id hash is within the tx range

Type Definitions

EpochIndex

Type representing domain epoch.

ExtrinsicsRoot

The Merkle root of all extrinsics included in a bundle.

OpaqueBundle

Bundle with opaque extrinsics.

OpaqueBundles

List of OpaqueBundle.

OperatorId

Type representing operator ID

OperatorPair

An operator authority keypair. Necessarily equivalent to the schnorrkel public key used in the main executor module. If that ever changes, then this must, too.

OperatorPublicKey

An operator authority identifier.

OperatorSignature

An operator authority signature.

ReceiptHash

The hash of a execution receipt.

RuntimeId

Type representing the runtime ID.

StakeWeight

Stake weight in the domain bundle election.