sp_domains::core_api

Trait DomainCoreApi

pub trait DomainCoreApi<Block: BlockT>: Core<Block> {

    // Provided methods
    fn extract_signer(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<ExtrinsicFor<Block>>,
    ) -> Result<Vec<(Option<AccountId>, Block::Extrinsic)>, ApiError> { ... }
    fn is_within_tx_range(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &Block::Extrinsic,
        bundle_vrf_hash: &U256,
        tx_range: &U256,
    ) -> Result<bool, ApiError> { ... }
    fn extract_signer_if_all_within_tx_range(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &Vec<Block::Extrinsic>,
        bundle_vrf_hash: &U256,
        tx_range: &U256,
    ) -> Result<Result<Vec<Option<AccountId>>, u32>, ApiError> { ... }
    fn initialize_block_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        header: &Block::Header,
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn apply_extrinsic_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: Block::Extrinsic,
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_set_code_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        code: Vec<u8>,
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_timestamp_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        moment: Moment,
    ) -> Result<Block::Extrinsic, ApiError> { ... }
    fn construct_consensus_chain_byte_fee_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        consensus_chain_byte_fee: Balance,
    ) -> Result<Block::Extrinsic, ApiError> { ... }
    fn construct_domain_update_chain_allowlist_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        updates: DomainAllowlistUpdates,
    ) -> Result<Block::Extrinsic, ApiError> { ... }
    fn is_inherent_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &Block::Extrinsic,
    ) -> Result<bool, ApiError> { ... }
    fn find_first_inherent_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: &Vec<Block::Extrinsic>,
    ) -> Result<Option<u32>, ApiError> { ... }
    fn check_extrinsics_and_do_pre_dispatch(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        uxts: Vec<Block::Extrinsic>,
        block_number: NumberFor<Block>,
        block_hash: Block::Hash,
    ) -> Result<Result<(), CheckExtrinsicsValidityError>, ApiError> { ... }
    fn decode_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        opaque_extrinsic: OpaqueExtrinsic,
    ) -> Result<Result<Block::Extrinsic, DecodeExtrinsicError>, ApiError> { ... }
    fn decode_extrinsics_prefix(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        opaque_extrinsics: Vec<OpaqueExtrinsic>,
    ) -> Result<Vec<Block::Extrinsic>, ApiError> { ... }
    fn extrinsic_era(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &Block::Extrinsic,
    ) -> Result<Option<Era>, ApiError> { ... }
    fn extrinsic_weight(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        ext: &Block::Extrinsic,
    ) -> Result<Weight, ApiError> { ... }
    fn extrinsics_weight(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        ext: &Vec<Block::Extrinsic>,
    ) -> Result<Weight, ApiError> { ... }
    fn block_fees(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<BlockFees<Balance>, ApiError> { ... }
    fn block_digest(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Digest, ApiError> { ... }
    fn block_weight(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Weight, ApiError> { ... }
    fn transfers(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Transfers<Balance>, ApiError> { ... }
    fn transfers_storage_key(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn block_fees_storage_key(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Vec<u8>, ApiError> { ... }

}

Base API that every domain runtime must implement.

Provided Methods

fn extract_signer( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<ExtrinsicFor<Block>>, ) -> Result<Vec<(Option<AccountId>, Block::Extrinsic)>, ApiError>

Extracts the optional signer per extrinsic.

fn is_within_tx_range( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &Block::Extrinsic, bundle_vrf_hash: &U256, tx_range: &U256, ) -> Result<bool, ApiError>

fn extract_signer_if_all_within_tx_range( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &Vec<Block::Extrinsic>, bundle_vrf_hash: &U256, tx_range: &U256, ) -> Result<Result<Vec<Option<AccountId>>, u32>, ApiError>

Extract signer of a given list of extrinsic if all of them are within the tx range, otherwise return the index if the first tx not in the tx range.

fn initialize_block_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, header: &Block::Header, ) -> Result<Vec<u8>, ApiError>

Returns the storage root after initializing the block.

fn apply_extrinsic_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: Block::Extrinsic, ) -> Result<Vec<u8>, ApiError>

Returns the storage root after applying the extrinsic.

fn construct_set_code_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, code: Vec<u8>, ) -> Result<Vec<u8>, ApiError>

Returns an encoded extrinsic aiming to upgrade the runtime using given code.

fn construct_timestamp_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, moment: Moment, ) -> Result<Block::Extrinsic, ApiError>

Returns an encoded extrinsic to set timestamp.

fn construct_consensus_chain_byte_fee_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, consensus_chain_byte_fee: Balance, ) -> Result<Block::Extrinsic, ApiError>

Returns an encoded extrinsic to set domain transaction byte fee.

fn construct_domain_update_chain_allowlist_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, updates: DomainAllowlistUpdates, ) -> Result<Block::Extrinsic, ApiError>

Returns an extrinsic to update chain allowlist.

fn is_inherent_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &Block::Extrinsic, ) -> Result<bool, ApiError>

Returns true if the extrinsic is an inherent extrinsic.

fn find_first_inherent_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: &Vec<Block::Extrinsic>, ) -> Result<Option<u32>, ApiError>

Find the first inherent extrinsic

fn check_extrinsics_and_do_pre_dispatch( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, uxts: Vec<Block::Extrinsic>, block_number: NumberFor<Block>, block_hash: Block::Hash, ) -> Result<Result<(), CheckExtrinsicsValidityError>, ApiError>

Checks the validity of array of extrinsics + pre_dispatch returning failure on first extrinsic that fails runtime call. IMPORTANT: Change CHECK_EXTRINSICS_AND_DO_PRE_DISPATCH_METHOD_NAME constant when this method name is changed

fn decode_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, opaque_extrinsic: OpaqueExtrinsic, ) -> Result<Result<Block::Extrinsic, DecodeExtrinsicError>, ApiError>

Decodes the domain specific extrinsic from the opaque extrinsic.

fn decode_extrinsics_prefix( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, opaque_extrinsics: Vec<OpaqueExtrinsic>, ) -> Result<Vec<Block::Extrinsic>, ApiError>

Decodes a list of domain extrinsics from opaque extrinsics, returning all extrinsics up to the first undecodable tx. (Or the full list if all are decodable.)

fn extrinsic_era( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &Block::Extrinsic, ) -> Result<Option<Era>, ApiError>

Returns extrinsic Era if present.

fn extrinsic_weight( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ext: &Block::Extrinsic, ) -> Result<Weight, ApiError>

Returns the extrinsic weight.

fn extrinsics_weight( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ext: &Vec<Block::Extrinsic>, ) -> Result<Weight, ApiError>

Returns the sum of a given set of extrinsics weight.

fn block_fees( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<BlockFees<Balance>, ApiError>

The accumulated transaction fee of all transactions included in the block.

fn block_digest( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Digest, ApiError>

Returns the block digest.

fn block_weight( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Weight, ApiError>

Returns the consumed weight of the block.

fn transfers( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Transfers<Balance>, ApiError>

Returns the transfers for this domain in the block.

fn transfers_storage_key( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Vec<u8>, ApiError>

Returns the storage key for the Transfers on Domain.

fn block_fees_storage_key( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Vec<u8>, ApiError>

Returns the storage key for the CollectedBlockFees on Domain.