Trait sp_domains::ExecutorApi

pub trait ExecutorApi<Block: BlockT, DomainHash: Encode + Decode>: Core<Block> {

    // Provided methods
    fn submit_bundle_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, Block::Hash, DomainHash>
    ) -> Result<(), ApiError> { ... }
    fn submit_bundle_unsigned_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, Block::Hash, DomainHash>
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }
    fn extract_system_bundles(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<(OpaqueBundles<Block, DomainHash>, SignedOpaqueBundles<Block, DomainHash>), ApiError> { ... }
    fn extract_system_bundles_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>
    ) -> Result<(OpaqueBundles<Block, DomainHash>, SignedOpaqueBundles<Block, DomainHash>), ApiError> { ... }
    fn extract_core_bundles(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<OpaqueBundles<Block, DomainHash>, ApiError> { ... }
    fn extract_core_bundles_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<OpaqueBundles<Block, DomainHash>, ApiError> { ... }
    fn extract_stored_bundle_hashes(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<H256>, ApiError> { ... }
    fn extract_stored_bundle_hashes_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Vec<H256>, ApiError> { ... }
    fn extract_receipts(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError> { ... }
    fn extract_receipts_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError> { ... }
    fn extract_fraud_proofs(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError> { ... }
    fn extract_fraud_proofs_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<Block::Extrinsic>,
        domain_id: DomainId
    ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError> { ... }
    fn extrinsics_shuffling_seed(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }
    fn extrinsics_shuffling_seed_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        header: Block::Header
    ) -> Result<Randomness, ApiError> { ... }
    fn system_domain_wasm_bundle(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Cow<'static, [u8]>, ApiError> { ... }
    fn system_domain_wasm_bundle_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Cow<'static, [u8]>, ApiError> { ... }
    fn head_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn head_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn system_domain_state_root_at(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        number: NumberFor<Block>,
        domain_hash: DomainHash
    ) -> Result<Option<Block::Hash>, ApiError> { ... }
    fn system_domain_state_root_at_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        number: NumberFor<Block>,
        domain_hash: DomainHash
    ) -> Result<Option<Block::Hash>, ApiError> { ... }

}

API necessary for executor pallet.

Provided Methods

fn submit_bundle_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, Block::Hash, DomainHash> ) -> Result<(), ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_bundle_unsigned_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, opaque_bundle: SignedOpaqueBundle<NumberFor<Block>, Block::Hash, DomainHash> ) -> Result<(), ApiError>

Submits the transaction bundle via an unsigned extrinsic.

fn submit_fraud_proof_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn submit_fraud_proof_unsigned_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

Submits the fraud proof via an unsigned extrinsic.

fn extract_system_bundles( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic> ) -> Result<(OpaqueBundles<Block, DomainHash>, SignedOpaqueBundles<Block, DomainHash>), ApiError>

Extract the system bundles from the given extrinsics.

fn extract_system_bundles_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic> ) -> Result<(OpaqueBundles<Block, DomainHash>, SignedOpaqueBundles<Block, DomainHash>), ApiError>

Extract the system bundles from the given extrinsics.

fn extract_core_bundles( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<OpaqueBundles<Block, DomainHash>, ApiError>

Extract the core bundles from the given extrinsics.

fn extract_core_bundles_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<OpaqueBundles<Block, DomainHash>, ApiError>

Extract the core bundles from the given extrinsics.

fn extract_stored_bundle_hashes( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<H256>, ApiError>

Extract the hashes of bundles stored in the block

fn extract_stored_bundle_hashes_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Vec<H256>, ApiError>

Extract the hashes of bundles stored in the block

fn extract_receipts( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_receipts_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<ExecutionReceipt<NumberFor<Block>, Block::Hash, DomainHash>>, ApiError>

Extract the receipts from the given extrinsics.

fn extract_fraud_proofs( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn extract_fraud_proofs_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<Block::Extrinsic>, domain_id: DomainId ) -> Result<Vec<FraudProof<NumberFor<Block>, Block::Hash>>, ApiError>

Extract the fraud proofs from the given extrinsics.

fn extrinsics_shuffling_seed( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, header: Block::Header ) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.

fn extrinsics_shuffling_seed_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, header: Block::Header ) -> Result<Randomness, ApiError>

Generates a randomness seed for extrinsics shuffling.

fn system_domain_wasm_bundle( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Cow<'static, [u8]>, ApiError>

WASM bundle for system domain runtime.

fn system_domain_wasm_bundle_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Cow<'static, [u8]>, ApiError>

WASM bundle for system domain runtime.

fn head_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn head_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

Returns the best execution chain number.

fn oldest_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn oldest_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

Returns the block number of oldest execution receipt.

fn maximum_receipt_drift( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.

fn maximum_receipt_drift_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

Returns the maximum receipt drift.

fn system_domain_state_root_at( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, number: NumberFor<Block>, domain_hash: DomainHash ) -> Result<Option<Block::Hash>, ApiError>

fn system_domain_state_root_at_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, number: NumberFor<Block>, domain_hash: DomainHash ) -> Result<Option<Block::Hash>, ApiError>